EUMJ: Fix sided conflicts
By adding sided proxies and doing some conditional seperation/trimming trough the FMLCommonHandler
the mod is now able to start on a physical server without crashing.

Some solutions/workarounds seen in this commit are subject to change as
the many of the fixes are lazy not very elegant.

Being very honest, this mode would require a complete rewrite that
treats common, client and server code seperately, and avoids mixing
counter sided code into one function, if not class.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>